+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Changed
+- Renamed the "show-status" action to "status"
+
+### Added
+- Add a logo for ga-spam-control
+
 ## v0.5.0
 
 ### Added