+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - Renamed the "remove-filters" action to "filters remove"
   - Renamed the "list-spam-domains" action to "domains list"
   - Renamed the "find-spam-domains" action to "domains find"
+- The spam-protection status calculation is now based on the number of spam domain names that are blocked by the existing filters - opposed to the number of filters which are up-to-date.
 
 ### Added
 - Add a logo for ga-spam-control